Move timing code into `getStaticProps()` and add 'Last updated' timestamp
